+++ CHANGELOG.md
@@ -10,6 +10,8 @@
   * found a way to display first loaded trip uid, approx duration and time of departure 152817699257cf3eee095d26c520ff40a1a074d8 ([15281769](https://github.com/CoVoCre/CoVoBETA/commit/152817699257cf3eee095d26c520ff40a1a074d8))
   * fetching and displaying of firestore trips working in /TestPage/Tests/CoVoTrips.js ! ([0a45452a](https://github.com/CoVoCre/CoVoBETA/commit/0a45452a17342383ffeee8d55c43c8a9eb393d54))
   * trying to get trip push to work in AddTrip - it now displays
+  * chose time is a string making possible to create a date and then call .getTime() to get timestamp in miliseconds, whereas firestore timestamp is just an object with .seconds and .miliseconds, so have to create time = new Date(.seconds\*1000) and then print date with just time
+  * adding PushTripToFirestore component to verify info and push new trip info, or to return error message
 
 * Working on AddTrip which should upload trip to cloud from Props